+++ project_specific/static/project_specific/js/calendar.js
@@ -400,309 +400,345 @@ function sendVacationDate(){
     
 }
 
-function createAcceptRejectButton (parentElement, acceptText='Accept', rejectText='Reject') {
-    let acceptButton = document.createElement("button")
-    let rejectButton = document.createElement("button")
-    acceptButton.innerText = acceptText
-    rejectButton.innerText = rejectText
-    acceptButton.setAttribute('onclick', 'acceptSwap(this)')
-    rejectButton.setAttribute('onclick', 'rejectSwap(this)')
-    parentElement.appendChild(acceptButton)
-    parentElement.appendChild(rejectButton)
-    return parentElement
-}
+// function createAcceptRejectButton (parentElement, acceptText='Request', rejectText='Reject') {
+//     let acceptButton = document.createElement("button")
+//     let rejectButton = document.createElement("button")
+//     acceptButton.innerText = acceptText
+//     rejectButton.innerText = rejectText
+//     acceptButton.setAttribute('onclick', 'applyRequest(this)')
+//     rejectButton.setAttribute('onclick', 'rejectSwap(this)')
+//     parentElement.appendChild(acceptButton)
+//     parentElement.appendChild(rejectButton)
+//     return parentElement
+// }
 
-function createCancelButton (parentElement) {
-    let cancelButton = document.createElement("button")
-    cancelButton.innerText = 'Cancel'
-    cancelButton.setAttribute('onclick', 'cancelSwapShift(this)')
-    parentElement.appendChild(cancelButton)
-    return parentElement
-}
+// function createCancelButton (parentElement) {
+//     let cancelButton = document.createElement("button")
+//     cancelButton.innerText = 'Cancel'
+//     cancelButton.setAttribute('onclick', 'cancelSwapShift(this)')
+//     parentElement.appendChild(cancelButton)
+//     return parentElement
+// }
 
-function cancelSwapShift (param) {
-    let parent_element = param.parentNode
-    let swap_shift_start = parent_element.dataset.shift_start
-    if (swap_shift_start) {
+// function cancelSwapShift (param) {
+//     let parent_element = param.parentNode
+//     let swap_shift_start = parent_element.dataset.shift_start
+//     if (swap_shift_start) {
 
-        let send_data = JSON.stringify({"action": "cancel", "data":swap_shift_start})
-        let csrftoken = getCookie('csrftoken')
-        $.ajax({
-            type: "POST",
-            url: "/main/swap/",
-            data: send_data,
-            headers: {
-                'X-CSRFToken': csrftoken
-            },
-            dataType: 'json',
-            success: function(result) {
-                if (result['status']) {
-                    alert('shift swap cancelled')
-                    fetchSwapResult()
-                } else {
-                    alert(result['error'])
-                }
-            },
-            contentType:'application/json'
-        })
-    }
-}
+//         let send_data = JSON.stringify({"action": "cancel", "data":swap_shift_start})
+//         let csrftoken = getCookie('csrftoken')
+//         $.ajax({
+//             type: "POST",
+//             url: "/main/swap/",
+//             data: send_data,
+//             headers: {
+//                 'X-CSRFToken': csrftoken
+//             },
+//             dataType: 'json',
+//             success: function(result) {
+//                 if (result['status']) {
+//                     alert('shift swap cancelled')
+//                     fetchSwapResult()
+//                 } else {
+//                     alert(result['error'])
+//                 }
+//             },
+//             contentType:'application/json'
+//         })
+//     }
+// }
 
-function acceptSwap (param) {
-    let parent_element = param.parentNode
-    let requester_shift_start = parent_element.parentNode.firstChild.dataset.shift_start
-    let acceptor_shift_start = parent_element.dataset.shift_start
-    let acceptor_employee_id = parent_element.dataset.employee_id
+// function applyRequest (param) {
+//     let parent_element = param.parentNode
+//     let requester_shift_start = parent_element.parentNode.firstChild.dataset.shift_start
+//     let acceptor_shift_start = parent_element.dataset.shift_start
+//     let acceptor_employee_id = parent_element.dataset.employee_id
 
-    let data = {'acceptor_shift_start': acceptor_shift_start,
-     'acceptor_employee_id': acceptor_employee_id,
-    'requester_shift_start': requester_shift_start}
+//     let data = {'acceptor_shift_start': acceptor_shift_start,
+//      'acceptor_employee_id': acceptor_employee_id,
+//     'requester_shift_start': requester_shift_start}
 
-    let send_data = JSON.stringify({"action": "request", "data":data})
-    let csrftoken = getCookie('csrftoken')
-    $.ajax({
-        type: "POST",
-        url: "/main/swap/request",
-        data: send_data,
-        headers: {
-            'X-CSRFToken': csrftoken
-        },
-        dataType: 'json',
-        success: function(result) {
-            if (result['status']) {
-                alert('Request sent')
-                fetchSwapResult()
-                fetchRequestResult()
+//     let send_data = JSON.stringify({"action": "request", "data":data})
+//     let csrftoken = getCookie('csrftoken')
+//     $.ajax({
+//         type: "POST",
+//         url: "/main/swap/request",
+//         data: send_data,
+//         headers: {
+//             'X-CSRFToken': csrftoken
+//         },
+//         dataType: 'json',
+//         success: function(result) {
+//             if (result['status']) {
+//                 alert('Request sent')
+//                 fetchSwapResult()
+//                 fetchRequestResult()
 
-            } else if (result['acceptor_error']) {
-                alert(result['acceptor_error'])
-            } else if (result['requester_error']) {
-                alert(result['requester_error'])
-            } else if (result['already_exist']) {
-                fetchSwapResult()
-                fetchRequestResult()
-                alert(`Request already exist for this shift`)
-            }
-        },
-        contentType:'application/json'
-    })
+//             } else if (result['acceptor_error']) {
+//                 alert(result['acceptor_error'])
+//             } else if (result['requester_error']) {
+//                 alert(result['requester_error'])
+//             } else if (result['already_exist']) {
+//                 fetchSwapResult()
+//                 fetchRequestResult()
+//                 alert(`Request already exist for this shift`)
+//             }
+//         },
+//         contentType:'application/json'
+//     })
 
-}
+// }
 
-function rejectSwap (param) {
-    let parent_element = param.parentNode
-    parent_element.remove()
-}
+// function rejectSwap (param) {
+//     // for applicant, rejecting swap result (clearing it for now)
+//     let parent_element = param.parentNode
+//     parent_element.remove()
+// }
 
-function fetchRequestResult() {
-    $.ajax({
-        type: "GET",
-        url: "/main/swap/request",
-        dataType: 'json',
-        success: function(response) {
-            $("#swapRequest-container").empty()
-            displayRequestResult(response)
-        },
-        contentType:'application/json'
-    })
-}
+// function fetchRequestResult() {
+//     // for applicant
+//     $.ajax({
+//         type: "GET",
+//         url: "/main/swap/request",
+//         dataType: 'json',
+//         success: function(response) {
+//             $("#swapRequest-container").empty()
+//             displayRequestResult(response)
+//         },
+//         contentType:'application/json'
+//     })
+// }
 
-function displayRequestResult(response) {
-    let swap_request_container = document.getElementById('swapRequest-container')
-    $('#swapRequest-container').empty()
-    for (num in response) {
-        let processing = response[num]
-        let shift_item = document.createElement('ul')
-        let applicant = document.createElement('li')
-        let receiver = document.createElement('li')
-        let status = ""
-        let cancelButton = document.createElement('button')
-        cancelButton.innerText = 'Cancel'
-        cancelButton.setAttribute('onclick', 'cancelRequest(this)')
-        shift_item.setAttribute('data-created_time', processing['created'])
+// function displayRequestResult(response) {
+//     // for applicant
+//     let swap_request_container = document.getElementById('swapRequest-container')
+//     $('#swapRequest-container').empty()
+//     for (num in response) {
+//         let processing = response[num]
+//         let shift_item = document.createElement('ul')
+//         let applicant = document.createElement('li')
+//         let receiver = document.createElement('li')
+//         let status = ""
+//         let cancelButton = document.createElement('button')
+//         cancelButton.innerText = 'Cancel'
+//         cancelButton.setAttribute('onclick', 'cancelRequest(this)')
+//         shift_item.setAttribute('data-created_time', processing['created'])
         
         
-        shift_item.setAttribute('data-applicant_shift_start', processing['applicant_shift_start'])
-        shift_item.setAttribute('data-applicant_shift_end', processing['applicant_shift_end'])
-        shift_item.setAttribute('data-receiver_shift_start', processing['receiver_shift_start'])
-        shift_item.setAttribute('data-receiver_shift_end', processing['receiver_shift_end'])
+//         shift_item.setAttribute('data-applicant_shift_start', processing['applicant_shift_start'])
+//         shift_item.setAttribute('data-applicant_shift_end', processing['applicant_shift_end'])
+//         shift_item.setAttribute('data-receiver_shift_start', processing['receiver_shift_start'])
+//         shift_item.setAttribute('data-receiver_shift_end', processing['receiver_shift_end'])
+//         shift_item.setAttribute('data-acceptor_employee_id', processing['receiver_employee_id'])
 
-        applicant.innerText = `Own Schedule ${processing['applicant_shift_start']} to ${processing['applicant_shift_end']}`
-        receiver.innerText = `Swap Schedule ${processing['receiver_shift_start']} to ${processing['receiver_shift_end']}`
-        if (processing['responded']) {
-            status = `${processing['accept']}`
-            if (processing['accept']) {
-                let acceptButton = document.createElement('button')
-                acceptButton.innerText = 'Accept'
-                acceptButton.setAttribute('onclick', 'acceptRequest(this)')
-                shift_item.innerHTML = status
-                shift_item.appendChild(acceptButton)
-                shift_item.appendChild(cancelButton)
-            } else {
-                shift_item.innerHTML = status
-                shift_item.appendChild(cancelButton)
-            }
-        } else {
-            status = 'Status: Not Responded'
-            shift_item.innerHTML = status
-            shift_item.appendChild(cancelButton)
-        }
+//         applicant.innerText = `Own Schedule ${processing['applicant_shift_start']} to ${processing['applicant_shift_end']}`
+//         receiver.innerText = `Swap Schedule ${processing['receiver_shift_start']} to ${processing['receiver_shift_end']}`
+//         if (processing['responded']) {
+//             status = `${processing['accept']}`
+//             if (processing['accept']) {
+//                 let acceptButton = document.createElement('button')
+//                 acceptButton.innerText = 'Accept'
+//                 acceptButton.setAttribute('onclick', 'finalizeSwap(this)')
+//                 shift_item.innerHTML = status
+//                 shift_item.appendChild(acceptButton)
+//                 shift_item.appendChild(cancelButton)
+//             } else {
+//                 shift_item.innerHTML = status
+//                 shift_item.appendChild(cancelButton)
+//             }
+//         } else {
+//             status = 'Status: Not Responded'
+//             shift_item.innerHTML = status
+//             shift_item.appendChild(cancelButton)
+//         }
         
-        shift_item.appendChild(applicant)
-        shift_item.appendChild(receiver)
+//         shift_item.appendChild(applicant)
+//         shift_item.appendChild(receiver)
         
-        swap_request_container.appendChild(shift_item)
-    }
-}
+//         swap_request_container.appendChild(shift_item)
+//     }
+// }
 
-function cancelRequest(param) {
-    let parent_element = param.parentNode
-    let created_time = parent_element.dataset.created_time
-    let requester_shift_start = parent_element.dataset.applicant_shift_start
-    console.log(created_time)
-    let data = {'created':created_time, 'requester_shift_start':requester_shift_start}
+// function cancelRequest(param) {
+//     // for applicant
+//     let parent_element = param.parentNode
+//     let created_time = parent_element.dataset.created_time
+//     let requester_shift_start = parent_element.dataset.applicant_shift_start
+//     let data = {'created':created_time, 'requester_shift_start':requester_shift_start}
 
-    let send_data = JSON.stringify({"action": "cancel", "data":data})
-    console.log(send_data)
-    let csrftoken = getCookie('csrftoken')
-    $.ajax({
-        type: "POST",
-        url: "/main/swap/request",
-        data: send_data,
-        headers: {
-            'X-CSRFToken': csrftoken
-        },
-        dataType: 'json',
-        success: function(result) {
-            if (result['status']) {
-                alert('Request cancelled')
-                fetchSwapResult()
-                fetchRequestResult()
-            } else {
-                alert(result['error_detail'])
-            }
-        },
-        contentType:'application/json'
-    })
-}
+//     let send_data = JSON.stringify({"action": "cancel", "data":data})
+//     let csrftoken = getCookie('csrftoken')
+//     $.ajax({
+//         type: "POST",
+//         url: "/main/swap/request",
+//         data: send_data,
+//         headers: {
+//             'X-CSRFToken': csrftoken
+//         },
+//         dataType: 'json',
+//         success: function(result) {
+//             if (result['status']) {
+//                 alert('Request cancelled')
+//                 fetchSwapResult()
+//                 fetchRequestResult()
+//             } else {
+//                 alert(result['error_detail'])
+//             }
+//         },
+//         contentType:'application/json'
+//     })
+// }
 
-function acceptRequest(param) {
+// function finalizeSwap(param) {
+//     // for applicant
+//     let parent_element = param.parentNode
+//     let created_time = parent_element.dataset.created_time
+//     let requester_shift_start = parent_element.dataset.applicant_shift_start
+//     let acceptor_shift_start = parent_element.dataset.acceptor_shift_start
+//     let acceptor_employee_id = parent_element.dataset.acceptor_employee_id
 
-}
 
-function fetchSwapResult () {
-    $.ajax({
-        type: "GET",
-        url: "/main/swap/",
-        dataType: 'json',
-        success: function(response) {
-            $("#swapResult-container").empty()
-            displaySwapResult(response)
-        },
-        contentType:'application/json'
-    })
-}
+//     let data = {'created':created_time, 'requester_shift_start':requester_shift_start,
+// 'acceptor_shift_start':acceptor_shift_start,'acceptor_employee_id':acceptor_employee_id}
 
-function displaySwapResult (result, new_info=false) {
-    let swapResultLists = document.getElementById('swapResult-container')
+//     let send_data = JSON.stringify({"action": "finalize", "data":data})
+//     let csrftoken = getCookie('csrftoken')
+//     $.ajax({
+//         type: "POST",
+//         url: "/main/swap/request",
+//         data: send_data,
+//         headers: {
+//             'X-CSRFToken': csrftoken
+//         },
+//         dataType: 'json',
+//         success: function(result) {
+//             if (result['status']) {
+//                 alert('Shift Swapped!')
+//                 fetchSwapResult()
+//                 fetchRequestResult()
+//             } else {
+//                 alert(result['error_detail'])
+//             }
+//         },
+//         contentType:'application/json'
+//     })
+// }
 
-    for (date in result) {
-        response = result[date]
-        let swapDateList = document.createElement('ul')
-        let swapDateListContainer = document.createElement('div')
-        swapDateListContainer.setAttribute('data-shift_start', `${date}`)
-        swapDateListContainer.innerText = date
-        swapDateListContainer = createCancelButton(swapDateListContainer)
+// function fetchSwapResult () {
+//     // for applicant
+//     $.ajax({
+//         type: "GET",
+//         url: "/main/swap/",
+//         dataType: 'json',
+//         success: function(response) {
+//             $("#swapResult-container").empty()
+//             displaySwapResult(response)
+//         },
+//         contentType:'application/json'
+//     })
+// }
 
-        swapDateListContainer.classList.add('flexbox')
-        swapDateList.appendChild(swapDateListContainer)
-        if (response['success']) {
-            let available_shifts = response['available_shifts']
-            if (available_shifts) {
+// function displaySwapResult (result, new_info=false) {
+//     // for applicant
+//     let swapResultLists = document.getElementById('swapResult-container')
+
+//     for (date in result) {
+//         response = result[date]
+//         let swapDateList = document.createElement('ul')
+//         let swapDateListContainer = document.createElement('div')
+//         swapDateListContainer.setAttribute('data-shift_start', `${date}`)
+//         swapDateListContainer.innerText = date
+//         swapDateListContainer = createCancelButton(swapDateListContainer)
+
+//         swapDateListContainer.classList.add('flexbox')
+//         swapDateList.appendChild(swapDateListContainer)
+//         if (response['success']) {
+//             let available_shifts = response['available_shifts']
+//             if (available_shifts) {
                 
-                for (detail in available_shifts) {
-                    let shift_detail = available_shifts[detail]
-                    let shift_start = shift_detail['shift_start']
-                    let shift_end = shift_detail['shift_end']
-                    let employee = shift_detail['employee']
-                    let employee_name =  `${employee['first_name']} ${employee['last_name']}`
+//                 for (detail in available_shifts) {
+//                     let shift_detail = available_shifts[detail]
+//                     let shift_start = shift_detail['shift_start']
+//                     let shift_end = shift_detail['shift_end']
+//                     let employee = shift_detail['employee']
+//                     let employee_name =  `${employee['first_name']} ${employee['last_name']}`
 
-                    let swaps = document.createElement('li')
-                    swaps.setAttribute("data-shift_start", `${shift_start}`)
-                    swaps.setAttribute("data-shift_end", `${shift_end}`)
-                    swaps.setAttribute("data-employee_id", `${employee['employee_id']}`)
+//                     let swaps = document.createElement('li')
+//                     swaps.setAttribute("data-shift_start", `${shift_start}`)
+//                     swaps.setAttribute("data-shift_end", `${shift_end}`)
+//                     swaps.setAttribute("data-employee_id", `${employee['employee_id']}`)
                     
-                    let info = document.createTextNode(`${employee_name} ${shift_start} to ${shift_end}`)
-                    swaps.appendChild(info)
-                    swaps = createAcceptRejectButton(swaps)
-                    swaps.classList.add('flexbox')
-                    swapDateList.appendChild(swaps)
-                }
-            } else if (response['available_people']) {
-                for (people of response['available_people']) {
-                    let swaps = document.createElement('li')
-                    swaps.setAttribute("data-shift_start", `${shift_start}`)
-                    swaps.setAttribute("data-shift_end", `${shift_end}`)
-                    swaps.setAttribute("data-employee_id", `${employee['employee_id']}`)
-                    let info = document.createTextNode(people)
-                    swaps.appendChild(info)
-                    swaps = createAcceptRejectButton(swaps)
-                    swaps.classList.add('flexbox')
-                    swapDateList.appendChild(swaps)
-                }
-            }
-        } else {
-            let swaps = document.createElement('li')
-            if (response['error']) {
-                if (new_info) {
-                    alert(response['error_detail'])
-                }
-                let info = document.createTextNode(response['error_detail'])
-                swaps.appendChild(info)
-                swapDateList.appendChild(swaps)
+//                     let info = document.createTextNode(`${employee_name} ${shift_start} to ${shift_end}`)
+//                     swaps.appendChild(info)
+//                     swaps = createAcceptRejectButton(swaps)
+//                     swaps.classList.add('flexbox')
+//                     swapDateList.appendChild(swaps)
+//                 }
+//             } else if (response['available_people']) {
+//                 for (people of response['available_people']) {
+//                     let swaps = document.createElement('li')
+//                     swaps.setAttribute("data-shift_start", `${shift_start}`)
+//                     swaps.setAttribute("data-shift_end", `${shift_end}`)
+//                     swaps.setAttribute("data-employee_id", `${employee['employee_id']}`)
+//                     let info = document.createTextNode(people)
+//                     swaps.appendChild(info)
+//                     swaps = createAcceptRejectButton(swaps)
+//                     swaps.classList.add('flexbox')
+//                     swapDateList.appendChild(swaps)
+//                 }
+//             }
+//         } else {
+//             let swaps = document.createElement('li')
+//             if (response['error']) {
+//                 if (new_info) {
+//                     alert(response['error_detail'])
+//                 }
+//                 let info = document.createTextNode(response['error_detail'])
+//                 swaps.appendChild(info)
+//                 swapDateList.appendChild(swaps)
             
-            } else {
-                let info = document.createTextNode('cannot find anyone to swap')
-                swaps.appendChild(info)
-                swapDateList.appendChild(swaps)
-            }
+//             } else {
+//                 let info = document.createTextNode('cannot find anyone to swap')
+//                 swaps.appendChild(info)
+//                 swapDateList.appendChild(swaps)
+//             }
             
-        }
+//         }
 
-        swapResultLists.appendChild(swapDateList)
-    }
-}
+//         swapResultLists.appendChild(swapDateList)
+//     }
+// }
 
-function fetchVacationResult () {
-    $.ajax({
-        type: "GET",
-        url: "/main/vacation/",
-        dataType: 'json',
-        success: function(response) {
-            displayVacationResult(response)
-        },
-        contentType:'application/json'
-    })
-}
+// function fetchVacationResult () {
+//     $.ajax({
+//         type: "GET",
+//         url: "/main/vacation/",
+//         dataType: 'json',
+//         success: function(response) {
+//             displayVacationResult(response)
+//         },
+//         contentType:'application/json'
+//     })
+// }
 
-function displayVacationResult (response) {
-    let vacationResultContainer = document.getElementById('vacationResult-container')
-    $("#vacationResult-container").empty()
-    for (date in response) {
-        let status = response[date]
-        let vacationList = document.createElement('ul')
-        let vacationDate = document.createTextNode(date)
-        vacationList.appendChild(vacationDate)
-        for (detail in status) {
-            let vacation_detail = document.createElement('li')
-            vacation_detail.innerText = `${detail}: ${status[detail]}`
-            vacationList.appendChild(vacation_detail)
-        }
-        vacationResultContainer.appendChild(vacationList)
-    }
-}
+// function displayVacationResult (response) {
+//     let vacationResultContainer = document.getElementById('vacationResult-container')
+//     $("#vacationResult-container").empty()
+//     for (date in response) {
+//         let status = response[date]
+//         let vacationList = document.createElement('ul')
+//         let vacationDate = document.createTextNode(date)
+//         vacationList.appendChild(vacationDate)
+//         for (detail in status) {
+//             let vacation_detail = document.createElement('li')
+//             vacation_detail.innerText = `${detail}: ${status[detail]}`
+//             vacationList.appendChild(vacation_detail)
+//         }
+//         vacationResultContainer.appendChild(vacationList)
+//     }
+// }
 
 buildCalendar (year, month)
-fetchVacationResult()
-fetchSwapResult()
-fetchRequestResult()
+// fetchVacationResult()
+// fetchSwapResult()
+// fetchRequestResult()
